# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeagueCreate": {
        "properties": {
          "competition_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Id"
          },
          "head_to_head_enabled": {
            "default": false,
            "title": "Head To Head Enabled",
            "type": "boolean"
          },
          "local_identity": {
            "default": "local-owner",
            "maxLength": 160,
            "minLength": 1,
            "title": "Local Identity",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/LeagueMode"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_display_name": {
            "default": "Allenatore",
            "maxLength": 160,
            "minLength": 1,
            "title": "Owner Display Name",
            "type": "string"
          },
          "season_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Id"
          },
          "team_name": {
            "default": "La mia rosa",
            "maxLength": 160,
            "minLength": 1,
            "title": "Team Name",
            "type": "string"
          },
          "timezone": {
            "default": "Europe/Rome",
            "title": "Timezone",
            "type": "string"
          },
          "total_credits": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "500",
            "title": "Total Credits"
          }
        },
        "required": [
          "name",
          "mode"
        ],
        "title": "LeagueCreate",
        "type": "object"
      },
...
      "LeaguePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeagueSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "LeaguePage",
        "type": "object"
      },
...
      "MarketImportRequest": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/ImportPlayer"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Players",
            "type": "array"
          },
          "replace_existing": {
            "default": true,
            "title": "Replace Existing",
            "type": "boolean"
          }
        },
        "required": [
          "players"
        ],
        "title": "MarketImportRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Probabilistic decision support for Italian Fantacalcio.",
    "title": "MarkGuardiola",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/leagues": {
      "get": {
        "operationId": "listLeagues",
        "parameters": [
          {
            "in": "query",
            "name": "local_identity",
            "required": false,
            "schema": {
              "default": "local-owner",
              "maxLength": 160,
              "minLength": 1,
              "title": "Local Identity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaguePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Leagues",
        "tags": [
          "leagues"
        ]
      },
      "post": {
        "operationId": "createLeague",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create League",
        "tags": [
          "leagues"
        ]
      }
    },
    "/api/v1/leagues/{league_id}": {
      "get": {
        "operationId": "getLeague",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read League",
        "tags": [
          "leagues"
        ]
      },
      "patch": {
        "operationId": "updateLeagueSettings",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueSettingsWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update League Settings",
        "tags": [
          "leagues"
        ]
      }
    },
    "/api/v1/leagues/{league_id}/market/import": {
      "post": {
        "operationId": "importMarket",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Market",
        "tags": [
          "leagues"
        ]
      }
    },
...
    "/api/v1/leagues/{league_id}/rosters": {
      "get": {
        "operationId": "listLeagueRosters",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RosterView"
                  },
                  "title": "Response Listleaguerosters",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rosters",
        "tags": [
          "leagues"
        ]
      }
    },
    "/api/v1/leagues/{league_id}/rosters/import": {
      "post": {
        "operationId": "importRoster",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RosterImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Roster",
        "tags": [
          "leagues"
        ]
      }
    },
    "/api/v1/leagues/{league_id}/rules": {
      "put": {
        "operationId": "replaceLeagueRules",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueRulesWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueRulesView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace League Rules",
        "tags": [
          "leagues"
        ]
      }
    },
    "/api/v1/leagues/{league_id}/teams/{fantasy_team_id}/budget": {
      "put": {
        "operationId": "updateFantasyTeamBudget",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fantasy_team_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Fantasy Team Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Budget",
        "tags": [
          "leagues"
        ]
      }
    },
...
  }
}